Added Royal Diamonds feature
I feel good about this code, but the requirements for this feature are
pretty weird. For example, Warren's map file currently doesn't contain
any diamonds! I talk about this in comments.

It does basically work, as you can see in a debugger running the test
html file.

I think we'll need to revisit this a few times, like if/when we add
support for items, because the diamonds may end up being a kind of item.

Signed-off-by: DanielBrewerPsu <brewer9@pdx.edu>