Added support for energy and current location
I went ahead and wrote the code into the webpage to display current location, energy, and whiffles.
Right now it is just below the buttons like in Warren's example, but I understand we might not put it there in the end.
Whiffles is showing up as undefined, and the buttons still move the location after the character is dead,
both things that seem to be issues within charModel.  I displayed the values with spans instead of text boxes, because,
frankly it looks less cheesy.

Signed-off-by: Nick Robinson <nro2@pdx.edu>