Tiles now contain objects
There was a small problem handling loading of tiles, this
had to be broke out similar to each model.

Also removed a few lines that were not needed in the setVisible function
while debugging this, now most functions call getTile which checks if a
tile exists to begin with.  initTileMeadow inits without setting to
visited by default. It seems to not make a difference either way as
tiles are generally only accessed with getTile when they are visible,
but in case they are accessed when not visible they wouldn't be set
to visible.

Signed-off-by: David Post <post@pdx.edu>