Put the Royal Diamonds in the raw map files
Previously, I had merely added the diamond object to the
Sample_Frugal.html (which is misspelled btw: Frupal vs Frugal). But that
wasn't exercising the whole loadMaps.py system or storing the diamonds
somewhere more permanent.

This checkin puts the diamonds in the maps/raw/default.map file and
includes a SampleFrugal.html based on that default.map.

Signed-off-by: DanielBrewerPsu <brewer9@pdx.edu>