Fixed error message about "messageDiv is null"
I think this was happening because controller.initGame, which calls
MessageView.redraw, appeared above the definition on the "message" div
in the html file. The Firefox browser does read this file top to bottom,
so "message" has not been define at the time that initGame was called.
I fixed this by moving controller.initGame in the file. Another approach
to this category of bug is to use onload and ensure that the whole file
is loaded before doing controller.initGame.

Signed-off-by: DanielBrewerPsu <brewer9@pdx.edu>