Sight
Added a member variable called sight that is now what sets sightRange in setVisible.  sight will be changed once the binoculars are added to the inventory.  I don't particularly care for setting the sight in mapModel but I believe this is the cleanest and easiest way to reveal more tiles.  Also tested this with sight = 2 and the tiles are revealed as expected.