Changed visRange to be default 1
Set this as a default 1 instead of loading it from the constructor. The loadmap
script would have needed updating, along with all maps in use, in order to add
something to the constructor. Seeing how the constructor is only really used for
loadMap.py, and there is a simple default, we are setting it in the constructor.

Signed-off-by: David Post <post@pdx.edu>