Updated gameView.html
I edited the "Messages to Read" to just say "Messages".  I also removed the bacon ipsum from the message div.  Doing this caused the
grey div boxes to not be lined up, so I edited the .css to increase the padding of the box, and now they line up.  It looks a little
better.

Signed-off-by: Nick Robinson <nro2@pdx.edu>