Added a value for Royal Diamonds
This fixes a recently introduced issue: When the player picks up the
royal diamonds, their value is listed as "undefined". Now, they are
listed as this colorful "999,999+" number. In testing, that value seems
to work elsewhere, because there's no code elsewhere that depends on
those values being numeric (like in a purchasing system).

Signed-off-by: DanielBrewerPsu <brewer9@pdx.edu>